requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
requests==2.21.0
selenium==3.141.0
soupsieve==1.8
tqdm==4.31.1
urllib3==1.24.1
wget==3.2
